requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
# Add this to your requirements.txt (if you're using one)
apache-airflow-providers-http
